improveMonsterMovement
I added so that you can turn of and turn on the flashlight with "F" and if you have the flashlight on the monster can see you quite far away, but if you have it turn of he cant se you so well. I also changed so that the monster runs to the last point he saw the player when the monster cant see the player any longer.